+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.5.0
 
 * Add `:year:`, `:month:`, `:day:`, `:hour:`, `:minute:` variables (#24)
+* Add last used menu entry
 
 # 1.4.4